sm8150-common: fod: Add option to keep the dimlayer always on
We are all well aware of how a special boy with an equally special FOD
OnePlus 7T (aka hotdogb) is, that's why I decided to keep its dimlayer
always active in lockscreen. This way there isn't even a glimpse of
flicker, at the expense of having a tinted lockscreen and a very bright
FOD circle.
But now StatiX supports two new sm8150 devices (OnePlus 7 and
OnePlus 7 Pro) which don't necessarily need an always-on dimlayer.
This commit guards everything with the flag "DIMLAYER", so the
maintainers are free to choose what's better for their devices by
setting `SOONG_CONFIG_ONEPLUS_MSMNILE_FOD_DIMLAYER` `true` or `false`
in the BoardConfig file of the device-specific tree.

Change-Id: Ib09281ea8eb54b783dc321776d553770c5a8c308